Solution - CALIBRATION and LIFTOFF detection
Solution commit - declaration, init, calibration and liftoff detection implemented.
This approach takes under consideration the mean before the liftoff and variance of the sensors.
For now this solution does not consider the case of outliers by the sensor, considering the liftoff upon the first sample that clearly is over the noise by a multiplication factor.
Next step is to detect apogee and landing.
Possible further [optional] implementations could be the detection of engine_off, due to decreased acc_x, the parachute deployment due to delta acceleration.