Graphics fix and apogee detection-oriented ones
Fixed the legend about the graphics (since it is an overall graphic about just the trend, I used a "normalized" barometric pressure but the formula was not on the legend.
Added some graphics related to apogee-detection. These 2 graphics takes the raw barometric pressures data without scaling/normalization/mean reduction. 
One of the two visualizes the trend of the barometric pressure which is not smooth but the mean is roughly quadratic.