Implemented Apogee detection - barometric based sol.
Implemented the apogee detection.
The solution is based on the barometric pressure tendency, after liftoff of decrease. 
Two "sample windows" are used for computing their mean and see if the tendency is of a minimum point. 
The data in apogee is really "dirty" so, possible implementations could use multiple data or outliers cleaning procedure.
So far, I just adjust the sample window width in a parametric way to avoid local minima to trigger apogee detection.